% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hier_clust.R
\name{hier_clust}
\alias{hier_clust}
\title{Implements simple linear regression by hand}
\usage{
hier_clust(dataset)
}
\arguments{
\item{dataset}{A data frame}
}
\value{
clusters
}
\description{
Implements simple linear regression by hand
}
